move EC algorithms to their own benchmark file
